# 8252848: Optimize small primitive arrayCopy operations through partial inlining using AVX-512 masked instructions