--- a/src/hotspot/cpu/s390/s390.ad
+++ b/src/hotspot/cpu/s390/s390.ad
@@ -278,8 +278,8 @@ alloc_class chunk2(
 // information in this architecture description.
 
 // 1) reg_class inline_cache_reg           (as defined in frame section)
-// 2) reg_class compiler_method_oop_reg    (as defined in frame section)
-// 2) reg_class interpreter_method_oop_reg (as defined in frame section)
+// 2) reg_class compiler_method_reg        (as defined in frame section)
+// 2) reg_class interpreter_method_reg     (as defined in frame section)
 // 3) reg_class stack_slots(/* one chunk of stack-based "registers" */)
 
 // Integer Register Classes
@@ -2468,10 +2468,10 @@ frame %{
   // interpreter_arg_ptr_reg(Z_R6);
 
   // Temporary in compiled entry-points
-  // compiler_method_oop_reg(Z_R1);//Z_R1_scratch
+  // compiler_method_reg(Z_R1);//Z_R1_scratch
 
-  // Method Oop Register when calling interpreter
-  interpreter_method_oop_reg(Z_R9);//Z_method
+  // Method Register when calling interpreter
+  interpreter_method_reg(Z_R9);//Z_method
 
   // Optional: name the operand used by cisc-spilling to access
   // [stack_pointer + offset].
@@ -3536,15 +3536,15 @@ operand inline_cache_regP(iRegP reg) %{
   interface(REG_INTER);
 %}
 
-operand compiler_method_oop_regP(iRegP reg) %{
-  constraint(ALLOC_IN_RC(z_r1_RegP)); // compiler_method_oop_reg
+operand compiler_method_regP(iRegP reg) %{
+  constraint(ALLOC_IN_RC(z_r1_RegP)); // compiler_method_reg
   match(reg);
   format %{ %}
   interface(REG_INTER);
 %}
 
-operand interpreter_method_oop_regP(iRegP reg) %{
-  constraint(ALLOC_IN_RC(z_r9_regP)); // interpreter_method_oop_reg
+operand interpreter_method_regP(iRegP reg) %{
+  constraint(ALLOC_IN_RC(z_r9_regP)); // interpreter_method_reg
   match(reg);
   format %{ %}
   interface(REG_INTER);