--- a/src/hotspot/cpu/zero/interpreterFrame_zero.hpp
+++ b/src/hotspot/cpu/zero/interpreterFrame_zero.hpp
@@ -57,8 +57,8 @@ class InterpreterFrame : public ZeroFrame {
  protected:
   enum Layout {
     istate_off = jf_header_words +
-      (align_up_(sizeof(BytecodeInterpreter),
-                 wordSize) >> LogBytesPerWord) - 1,
+      (align_up(sizeof(BytecodeInterpreter),
+                wordSize) >> LogBytesPerWord) - 1,
     header_words
   };