… src/hotspot/cpu/ppc/vm_version_ppc.cpp
… src/hotspot/share/adlc/formsopt.hpp
… src/hotspot/share/adlc/formssel.cpp
… src/hotspot/share/adlc/output_c.cpp
… src/hotspot/share/ci/ciEnv.cpp
… src/hotspot/share/ci/ciField.cpp
… src/hotspot/share/ci/ciUtilities.inline.hpp
… src/hotspot/share/classfile/systemDictionaryShared.cpp
… src/hotspot/share/classfile/systemDictionaryShared.hpp
… src/hotspot/share/compiler/compileBroker.cpp
… src/hotspot/share/compiler/compilerEvent.cpp
… src/hotspot/share/compiler/compilerEvent.hpp
… src/hotspot/share/gc/epsilon/epsilonArguments.cpp
… src/hotspot/share/gc/epsilon/epsilonBarrierSet.cpp
… src/hotspot/share/gc/epsilon/epsilonHeap.cpp
… src/hotspot/share/gc/epsilon/epsilonMonitoringSupport.cpp
… src/hotspot/share/gc/g1/g1Analytics.cpp
… src/hotspot/share/gc/g1/g1ParScanThreadState.cpp
… src/hotspot/share/gc/g1/g1ParScanThreadState.hpp
… src/hotspot/share/gc/g1/g1Policy.cpp
… src/hotspot/share/gc/g1/g1Policy.hpp
… src/hotspot/share/gc/shared/collectedHeap.cpp
… src/hotspot/share/gc/shared/partialArrayTaskStepper.cpp
… src/hotspot/share/gc/shared/partialArrayTaskStepper.hpp
… src/hotspot/share/gc/shared/partialArrayTaskStepper.inline.hpp
… src/hotspot/share/gc/shared/taskqueue.hpp
… src/hotspot/share/gc/shenandoah/shenandoahRootProcessor.cpp
… src/hotspot/share/gc/shenandoah/shenandoahStrDedupQueue.cpp
… src/hotspot/share/gc/shenandoah/shenandoahStrDedupQueue.hpp
… src/hotspot/share/gc/z/zForwardingEntry.hpp