@dholmes-ora 2020-09-14
/csr needed

Adding a new product flag requires a CSR request to  be filed.