@jatin-bhateja 2020-09-14
> /csr needed
> 
> Adding a new product flag requires a CSR request to be filed.

@dholmes-ora ,  with https://github.com/openjdk/jdk/commit/5144190e  there has been a clean up of options and product options now accept  DIAGNOSTIC as an additional parameter.  Newly added flag is a DIAGNOSTIC flag.