@dholmes-ora 2020-09-15
> 
> 
> > /csr needed
> > Adding a new product flag requires a CSR request to be filed.
> 
> @dholmes-ora , with [5144190](https://github.com/openjdk/jdk/commit/5144190e) there has been a clean up of options and product options now accept DIAGNOSTIC as an additional parameter. Newly added flag is a DIAGNOSTIC flag.

Apologies for that. Yes I got caught out by the new format.

/csr notneeded