@dholmes-ora 2020-09-15
/csr unneeded